+++ data/src/main/java/com/BySandS/testsandquizes/data/test/storage/dao/DaoQuestionStorageImpl.kt
@@ -10,7 +10,7 @@ class DaoQuestionStorageImpl(context: Context) : QuestionStorage {
     private val testDao = AppDatabase.getInstance(context).questionDao()
     override fun getQuestionsList(
         difficultyId: Long,
-        quantityOfQuestions: Long
+        quantityOfQuestions: Int
     ): List<QuestionModelDb> {
         return testDao.getQuestionsListByDifficulty(
             difficultyId = difficultyId,
