+++ data/src/main/java/com/BySandS/testsandquizes/data/test/storage/dao/interfaceDao/DaoQuestionRu.kt
@@ -9,5 +9,5 @@ interface DaoQuestionRu {
     @Query("SELECT * FROM  question " +
             "WHERE question.difficulty_id = :difficultyId " +
             "ORDER BY RANDOM() LIMIT :quantityOfQuestions")
-    fun getQuestionsListByDifficulty(difficultyId: Long, quantityOfQuestions: Long): List<QuestionModelDb>
+    fun getQuestionsListByDifficulty(difficultyId: Long, quantityOfQuestions: Int): List<QuestionModelDb>
 }
